Provide access to Space Track v2 REST interface, which is currently in beta. This involves adding attributes domain_space_track_rest, rest_session_cookie, and space_track_version.
Fix problem with Astro::SpaceTrack file() method passing its options
hash to the retrieve() method as an OID. This was probably introduced in
version 0.020.

Bump Test::More requirement from 0.88 to 0.96 so I can do subtests.

Add method content_interface, which returns the version of the Space
Track interface use to fetch the data. If Space Track was not involved,
returns undef.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-SpaceTrack/trunk@3294 f6fd3633-1913-47c2-bad4-372d75c948ec