I figured out I could implement epoch selection of TLEs in the Space Track version 2 REST interface by using its range operator on the desired epoch dates. So I did, and this now works.
Rather than hard-code a single space as '%20', run the REST arguments
through URI::Escape::uri_escape(). Add URI::Escape to the requirements.

Tweak the way _get_rest() arguments are processed internally. The first
argument after the invocant is no longer handled separately, but
absorbed into the @args array.

When _get_rest() is called with a hash, the hash is unrolled into a list
by sorting the keys and running them through a map{}. This is an attempt
to normalize the order of the arguments, since they get passed to
_dump_request() as an array.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-SpaceTrack/trunk@3324 f6fd3633-1913-47c2-bad4-372d75c948ec