Simplify the Space Track v2 search code.
Retrieve multiple OIDs in a single call to the Space Track interface.
This involves a custom unsafe characters list to
URI::Escape::escape_uri, that adds comma to the RFC3986 list, since it
appears Space Track does not accept a percent-encoded comma.

Pass the entire escaped URL to the request dumper.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-SpaceTrack/trunk@3357 f6fd3633-1913-47c2-bad4-372d75c948ec