Have the output of names( 'spacetrack' ) depend on the value of the space_track_version attribute.
Rough out an implementation for _spacetrack_v2(), based on queries to
the REST interface. It is not possible to duplicate all possible v1 bulk
data catalogs this way, but a few are possible in theory. The code is
currenyly commented out because it is very slow, and /class/tle queries
with large numbers of OIDs seem to result in HTTP error 500.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-SpaceTrack/trunk@3398 f6fd3633-1913-47c2-bad4-372d75c948ec