Fix the joining of JSON data when one of the lists is empty.
Have the celestrak() and file() methods supply common names from their
sources if the with_name attribute is true and the space_track_version
attribute is 2. This functionality will be retracted when (and if) the
Space Track REST interface becomes capable of supplying NASA-format
TLEs.

Document the fact that Space Track has deprecated bulk data downloads,
and plans to remove them in October 2012. This will break the
spacetrack() method.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-SpaceTrack/trunk@3446 f6fd3633-1913-47c2-bad4-372d75c948ec