Add the 'full' and 'payloads' catalogs to the REST version of the spacetrack() method.
Add the 'time' pseudo-command to the shell() method. Its arguments
become the command to run, and its effect is to cause that command to be
timed.

Enhance the REST version of the spacetrack() method to allow the
specification of queries that only use class 'tle_current' if common
names are not required. Take advantage of this enhancement for the
'full' and 'geosynchronous' catalogs.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-SpaceTrack/trunk@3479 f6fd3633-1913-47c2-bad4-372d75c948ec