Validate options passed as hash ref
Formerly, options passed as a hash reference were not validated. This
change causes a warning if any unrecognized keys are found. The problem
with simply ignoring them was that the caller gets no feedback on why an
option is not honored.

Environment variable SPACETRACK_SKIP_OPTION_HASH_VALIDATION can be set
to suppress the warning.

The plan is to make them fatal eventually, and to ignore the environment
variable.