Add -file option to amsat() and mccants()
Centralized code for a GET from a URL without authentication. If this
code gets a 'file' argument, the GET is done with the if_modified_since
header set to the modification date of the file. If the response code is
HTTP_NOT_MODIFIED, we fill in the content from the file, otherwise we
update the file from the (possibly post-processed) content. Either way
the content of the returned HTTP::Response object and the file end up
being the same.

Rewrote both amsat() and mccants() to take a -file option and go through
the centralized code.

Added cache_hit() and corresponding pragma, which is set (if at all) by
the centralized code, so I can verify whether the data came from the net
or the file. Refacored pragma access code while I was about it, since I
now had four methods that were the same but for the name of the pragma
accessed.