Fix problem detecting Space Track login failure.
This was specific to version 2 of the interface, but since this is the
only one for now, that's moot. You couldn't access the data, but login()
was returning success.

Apparently version 1 of the interface didn't return a cookie on login
failure, but version 2 does. On success version 2 returns status 200
content ''. On failure it still returns status 200, but content
'{"Login":"Failure"}', which looks like JSON to me. But it also returns
a cookie, which apparently is not valid.