Try to sanitize Zip data for mccants() method.
I'm getting failures under older Windows (5.2 and earlier) because
Archive::Zip claims the handle I gave it is not seekable. This update is
a two-liner (plus a gob of comments) that replaces
 open my $fh, '+<', \( $resp->content() )
with
 my $content = $resp->content();
 open my $fh, '+<', \$comment;
in the hope that the failure is due to some strangeness in the way
Windows implements HTTP::Response.

Actually, Windows is not the only culprit -- I have a lone instance of
teh same failure under FreeBSD 9.2-release. The failure had a bunch of
Acme::Override::INET replaced IO::Socket::INET with IO::Socket::IP
messages, but so did at least one of the FreeBSD passes, so I presume
for the moment that's not an issue.