Get rid of ugly warning when searching Space Track
The problem was that the retrieval options -start_epoch and -end_epoch
were being processed into times when the search_*() arguments were
parsed. This was falling afoul of the "no unknown keys" logic when the
options were passed to retrieve() as a hash reference. This was handled
by deferring this processing, since the search_*() methods do not need
the information.