Support the various Space Track data formats
This involves providing a -format option, with values corresponding to
the Space Track formats, plus 'legacy' (the default) to provide what is
essentially the Version 1 data format.

The old -json option is equivalent to -format json, and it is illegal to
combine -format and -json if -format specifies any format other than
json.