Prepare for removal of RCSVALUE
Space Track has announced the removal of the RCSVALUE key from their
satcat data, replacing it with RCS_SIZE (small, medium, large). This
update is mostly documentation, but it does include a couple lines
bypassing the storage of a satcat RCSVALUE value in the TLE data if the
RCSVALUE key is non-existent (as opposed to undef).