No -last5 unless Space Track class is 'tle_latest'
Thanks to Vidar Tyldum for finding the problem and running it to ground.

The symptom was that if more than one OID was specified for a retrieve()
AND the -since_file option was specified, only one OID was being
returned.

The aooarant cause was that if the Space Track query class 'tle' was
used, then information from -last5 (even if defaulted!) could leak into
the REST query and cause anomalous results. This turns out to affect not
only the retrieve() method, but the search_*() methods if any opyion
(such as -status) causes them not to use class tle_latest to retrueve
the TLE data.

The solution is to remove the REST equivalent of -last5 from the query
unless it was going against class tle_latest.