Better HTTP request/response dumping
Add DUMP_RESPONSE_TRUNCATED, 0x20, which truncates the content of the
response to 61 bytes (64 counting the elipsis that replaces the
truncated stuff). This is otherwise the same as DUMP_RESPONSE.

Add optional second argument (third, counting the invocant) to
_dump_request() and __dump_response(), which specifies an alternate
message identifying the dump. The default is the old message.

Dump response object in _get_from_net() as it comes out of HTTP, as well
as the object as it is just before it is returned.