Better test output in t/query_mcceants.t
The issue here is that pretty regularly the test seems to say that Mike
McCants' server is ignoring the If-Modified-Since header. The thing here
was to modify the semantics of the SPACETRACK_TEST_CACHE environment
variable, so that it can be used to turn on tracing. This means that 0
now specifies testing the cache logic where before it did not.

Investigation has lead to the conclusion the the server in question
rarely if ever honors the If-Modified-Since header, so the relevant test
(after all that work) is now TODO.