Correct generation of Sladen Iridum status
This code scrapes the relevant web page. An omitted right parenthesis at
the end of a parenthetical comment on Plane 2 in the November 15 2017
update caused plane 3 to be dropped from the output. Because Rod's
parenthetical comments seem never to span line breaks, I am now ending
comments at line breaks if a right parenthesis is not seen.

This commit also explicitly codes for the convention of using slashes to
separate the numbers of satellites sharing the same slot. So far this
only occurs in Plane 4, and this change does not seem to change the
output, since in all cases the sharing is an Iridium Classic and Iridium
Next. Call it defensive programming.