Make 'decayed' internal status override anything.
I am not sure why it was not this way anyway; the line that did this was
commented out. I can only assume I commented it out to do some test and
then forgot to re-enable it.

I commit with failures in xt/author/iridium_status.t because when Kelso
removed Iridium 8 from his iridium.txt he also dropped the status flags.