Add Celestrak catalogs 'active' and 'analyst'.
I have no idea what these actually represent.

Correlation of iridium.txt and active.txt shows that an Iridium
satellite appears in the latter if and only if its status in the former
is '[+]' (in-service), '[S]' (spare), or '[P]' (partly operational. So I
presume presence in active.txt means the satellite is still at least
partly functional. Where Dr. Kelso gets this information I have no idea.

The contents of analyst.txt all have common name 'UNKNOWN', and OIDs in
the range 81014 to 89496, with gaps. The couple I spot-checked do NOT
appear in the Space Track catalog.